Convert usecase to TypeScript
