+++ examples/todomvc-typescript/src/usecase/FilterTodoList.ts
@@ -1,6 +1,6 @@
 "use strict";
 import { UseCase } from "almin";
-import { FilterType } from "../store/TodoStore/TodoState";
+import { FilterTypes } from "../store/TodoStore/TodoState";
 export class FilterTodoListFactory {
     static create() {
         return new FilterTodoListUseCase();
@@ -8,7 +8,7 @@ export class FilterTodoListFactory {
 }
 
 export class FilterTodoListUseCase extends UseCase {
-    execute(filterType: FilterType) {
+    execute(filterType: FilterTypes) {
         this.dispatch({
             type: "FilterTodoListUseCase",
             filterType
