+++ examples/todomvc-typescript/src/usecase/RemoveAllCompletedItems.ts
@@ -1,26 +1,29 @@
 "use strict";
 import { UseCase } from "almin";
 import todoListRepository, { TodoListRepository } from "../infra/TodoListRepository";
-export class ToggleTodoItemFactory {
+
+export class RemoveTodoItemFactory {
     static create() {
-        return new ToggleTodoItemUseCase({
+        return new RemoveTodoItemUseCase({
             todoListRepository
         });
     }
 }
 
-export class ToggleTodoItemUseCase extends UseCase {
+export class RemoveTodoItemUseCase extends UseCase {
+    private todoListRepository: TodoListRepository;
+
     /**
      * @param {TodoListRepository} todoListRepository
      */
-    constructor({ todoListRepository }) {
+    constructor({ todoListRepository }: { todoListRepository: TodoListRepository }) {
         super();
         this.todoListRepository = todoListRepository;
     }
 
-    execute(itemId) {
+    execute() {
         const todoList = this.todoListRepository.lastUsed();
-        todoList.toggleComplete(itemId);
+        todoList.removeAllCompletedItems();
         this.todoListRepository.save(todoList);
     }
 }
